+++ src/framework/piral-core/src/components/index.ts
@@ -2,6 +2,7 @@ export * from './components';
 export * from './DefaultErrorInfo';
 export * from './DefaultLayout';
 export * from './DefaultLoader';
+export * from './DefaultRouter';
 export * from './ErrorBoundary';
 export * from './ExtensionSlot';
 export * from './Mediator';