+++ src/tooling/piral-cli/src/common/envs.ts
@@ -37,6 +37,10 @@ export function setStandardEnvs(options: StandardEnvProps) {
     delete process.env.DEBUG_PILET;
   }
 
+  if (options.publicPath) {
+    process.env.PIRAL_PUBLIC_PATH = options.publicPath;
+  }
+
   if (options.production) {
     process.env.NODE_ENV = 'production';
   } else if (!process.env.NODE_ENV) {