+++ src/tooling/piral-cli/src/injectors/piral.ts
@@ -22,6 +22,8 @@ export default class PiralInjector implements KrasInjector {
     const api = '/$events';
     const cbs = {};
 
+    core.setMaxListeners(16);
+
     core.on('user-connected', (e) => {
       if (e.target === '*' && e.url === api.substr(1)) {
         cbs[e.id] = (msg: string) => e.ws.send(msg);